- Keep all damage factors in the table, even those that are 1.0. - 256 is a valid pain chance, so clamp to that, not 255.
SVN r2237 (trunk)

+++ src/info.h
@@ -158,7 +158,7 @@ struct FPlayerColorSet
 };
 
 typedef TMap<FName, fixed_t> DmgFactors;
-typedef TMap<FName, BYTE> PainChanceList;
+typedef TMap<FName, int> PainChanceList;
 typedef TMap<int, FPlayerColorSet> FPlayerColorSetMap;
 
 struct FActorInfo
